Added publish date/draft mode, facebook like/share/comments

+++ blog/posts/views.py
@@ -1,13 +1,17 @@
-from django.shortcuts import render, get_object_or_404, redirect
+from django.shortcuts import render, get_object_or_404, redirect, Http404
 from django.http import HttpResponse, HttpResponseRedirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.contrib import messages
+from django.utils import timezone
 
 from .models import Post
 from .forms import PostForm
 
 def posts_list(request):
-    queryset_list = Post.objects.all().order_by("-timestamp")
+    today = timezone.now().date()
+    queryset_list = Post.objects.active().order_by("-timestamp")
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Post.objects.all().order_by("-timestamp")
     paginator = Paginator(queryset_list, 5)
     page = request.GET.get('page')
 
@@ -20,12 +24,16 @@ def posts_list(request):
 
     context = {
         "title": "List",
-        "object_list": queryset
+        "object_list": queryset,
+        "today": today
     }
     return render(request, "post_list.html", context)
 
 def post_detail(request, slug=None):
     instance = get_object_or_404(Post, slug=slug)
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or request.user.is_superuser:
+            raise Http404
     context = {
         "title": instance.title,
         "instance": instance
@@ -33,10 +41,16 @@ def post_detail(request, slug=None):
     return render(request, "post_detail.html", context)
 
 def post_create(request):
-    form = PostForm(request.POST or None, request.FILES or None)
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
 
+    if not request.user.is_authenticated():
+        raise Http404
+
+    form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Post created.")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -46,8 +60,11 @@ def post_create(request):
     }
     return render(request, "post_form.html", context)
 
-def post_update(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_update(request, slug=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
+
+    instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 
     if form.is_valid():
@@ -64,6 +81,9 @@ def post_update(request, id=None):
     return render(request, "post_form.html", context)
 
 def post_delete(request, id=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
+
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, "Successfully deleted.")
